+++ src/fn/wait.js
@@ -1,0 +1,10 @@
+/**
+ * Wait
+ * @param {number} timeout - Time in ms
+ * @return {Promise<void>} - Delays execution for the given ms
+ */
+module.exports = function wait( timeout ) {
+    return new Promise( ( resolve ) => {
+        setTimeout( resolve, timeout );
+    } );
+};