+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.5.2
+ - Added *Progress* and *Timer* classes.
+ - Added *callback* and *wait* functions.
+
 ## 0.5.1
  - Docs and added *FsInterface.isDir* method.